fix: add missing integer-hours migration and model changes
Migration c6d7e8f9a0b1 (already committed) declares
down_revision=b5c6d7e8f9a0, but that migration and the matching
model/schema changes (hours Numeric -> Integer) were never committed,
breaking 'alembic upgrade head' on fresh clones.

+++ backend/app/models/timesheet_entries.py
@@ -1,10 +1,9 @@
 from __future__ import annotations
 
 from datetime import date
-from decimal import Decimal
 from typing import TYPE_CHECKING
 
-from sqlalchemy import CheckConstraint, Date, ForeignKey, Index, Numeric, UniqueConstraint, func
+from sqlalchemy import CheckConstraint, Date, ForeignKey, Index, Integer, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -21,7 +20,7 @@ class TimesheetEntry(Base):
     employee_id: Mapped[int] = mapped_column(ForeignKey("employees.id"), index=True, nullable=False)
     work_date: Mapped[date] = mapped_column(Date, index=True, nullable=False)
     company_id: Mapped[int] = mapped_column(ForeignKey("companies.id"), nullable=False)
-    hours: Mapped[Decimal] = mapped_column(Numeric(4, 2), nullable=False)
+    hours: Mapped[int] = mapped_column(Integer, nullable=False)
 
     created_at: Mapped[str] = mapped_column(server_default=func.now())
     updated_at: Mapped[str] = mapped_column(server_default=func.now(), onupdate=func.now())
